Move test_st_plotly_chart to plotly_chart_test.py file
