+++ lib/tests/streamlit/streamlit_test.py
@@ -21,7 +21,6 @@ import subprocess
 import sys
 import tempfile
 import unittest
-from unittest.mock import patch
 
 import numpy as np
 import pandas as pd
@@ -177,57 +176,6 @@ class StreamlitTest(unittest.TestCase):
 
 class StreamlitAPITest(DeltaGeneratorTestCase):
     """Test Public Streamlit Public APIs."""
-
-    def test_st_plotly_chart_simple(self):
-        """Test st.plotly_chart."""
-        import plotly.graph_objs as go
-
-        trace0 = go.Scatter(x=[1, 2, 3, 4], y=[10, 15, 13, 17])
-
-        data = [trace0]
-
-        st.plotly_chart(data)
-
-        el = self.get_delta_from_queue().new_element
-        self.assertEqual(el.plotly_chart.HasField("url"), False)
-        self.assertNotEqual(el.plotly_chart.figure.spec, "")
-        self.assertNotEqual(el.plotly_chart.figure.config, "")
-        self.assertEqual(el.plotly_chart.use_container_width, False)
-
-    def test_st_plotly_chart_use_container_width_true(self):
-        """Test st.plotly_chart."""
-        import plotly.graph_objs as go
-
-        trace0 = go.Scatter(x=[1, 2, 3, 4], y=[10, 15, 13, 17])
-
-        data = [trace0]
-
-        st.plotly_chart(data, use_container_width=True)
-
-        el = self.get_delta_from_queue().new_element
-        self.assertEqual(el.plotly_chart.HasField("url"), False)
-        self.assertNotEqual(el.plotly_chart.figure.spec, "")
-        self.assertNotEqual(el.plotly_chart.figure.config, "")
-        self.assertEqual(el.plotly_chart.use_container_width, True)
-
-    def test_st_plotly_chart_sharing(self):
-        """Test st.plotly_chart when sending data to Plotly's service."""
-        import plotly.graph_objs as go
-
-        trace0 = go.Scatter(x=[1, 2, 3, 4], y=[10, 15, 13, 17])
-
-        data = [trace0]
-
-        with patch(
-            "streamlit.elements.plotly_chart." "_plot_to_url_or_load_cached_url"
-        ) as plot_patch:
-            plot_patch.return_value = "the_url"
-            st.plotly_chart(data, sharing="public")
-
-        el = self.get_delta_from_queue().new_element
-        self.assertEqual(el.plotly_chart.HasField("figure"), False)
-        self.assertNotEqual(el.plotly_chart.url, "the_url")
-        self.assertEqual(el.plotly_chart.use_container_width, False)
 
     def test_st_legacy_table(self):
         """Test st._legacy_table."""
